Fix supabase, database works
