+++ package.json
@@ -7,9 +7,10 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "server": "nodemon backend/server.cjs",
-    "single-server": "node backend/server.cjs",
-    "serve": "concurrently \"next dev\" \"npm run server\""
+    "server": "nodemon server.mjs",
+    "single-server": "node server.mjs",
+    "serve": "concurrently \"next dev\" \"npm run server\"",
+    "pages:build": "npx @cloudflare/next-on-pages"
   },
   "dependencies": {
     "@mendable/firecrawl-js": "^0.0.21",
@@ -37,6 +38,7 @@
     "express": "^4.19.2",
     "jquery": "^3.7.1",
     "lucide-react": "^0.381.0",
+    "module-alias": "^2.2.3",
     "next": "14.2.3",
     "react": "^18",
     "react-day-picker": "^8.10.1",
@@ -46,6 +48,7 @@
     "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
+    "@cloudflare/next-on-pages": "^1.11.3",
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
     "@types/jquery": "^3.5.30",
@@ -58,5 +61,8 @@
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
   },
-  "type": "module"
+  "type": "commonjs",
+  "_moduleAliases": {
+    "@": "proteinscraper"
+  }
 }
