+++ tsconfig.json
@@ -23,6 +23,6 @@
     
   },
   
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "backend/server.cjs", "cheerio.ts"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "server.mjs", "cheerio.ts"],
   "exclude": ["node_modules"]
 }
